+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentController.scala
@@ -882,9 +882,7 @@ class GridAgentController(
         emInput,
         modelConfiguration,
         outputConfig,
-        maybeControllingEm
-          .map(_ => "PRIORITIZED")
-          .getOrElse("PROPORTIONAL"),
+        emInput.getControlStrategy,
         simulationStartDate,
         maybeControllingEm.toRight(
           environmentRefs.scheduler