+++ src/main/scala/edu/ie3/simona/service/ev/ExtEvDataService.scala
@@ -6,7 +6,7 @@
 
 package edu.ie3.simona.service.ev
 
-import edu.ie3.simona.api.data.ev.ExtEvData
+import edu.ie3.simona.api.data.ev.ExtEvDataConnection
 import edu.ie3.simona.api.data.ev.model.EvModel
 import edu.ie3.simona.api.data.ev.ontology._
 import edu.ie3.simona.api.data.ontology.DataMessageFromExt
@@ -52,7 +52,7 @@ import scala.util.{Failure, Success, Try}
 object ExtEvDataService {
 
   final case class ExtEvStateData(
-      extEvData: ExtEvData,
+      extEvData: ExtEvDataConnection,
       uuidToActorRef: Map[UUID, ClassicRef] = Map.empty[UUID, ClassicRef],
       extEvMessage: Option[EvDataMessageFromExt] = None,
       freeLots: ReceiveDataMap[UUID, Int] = ReceiveDataMap.empty,
@@ -61,7 +61,7 @@ object ExtEvDataService {
   ) extends ServiceBaseStateData
 
   final case class InitExtEvData(
-      extEvData: ExtEvData
+      extEvData: ExtEvDataConnection
   ) extends InitializeServiceStateData
 
   def adapter(evService: ActorRef[EvMessage]): Behavior[DataMessageFromExt] =