+++ src/test/scala/edu/ie3/simona/agent/participant/EvcsAgentModelCalculationSpec.scala
@@ -1383,9 +1383,9 @@ class EvcsAgentModelCalculationSpec
       resultListener.expectMsgPF() { case FlexOptionsResultEvent(flexResult) =>
         flexResult.getInputModel shouldBe evcsInputModelQv.getUuid
         flexResult.getTime shouldBe 900.toDateTime
-        flexResult.getpRef should beEquivalentTo(ev900.unwrap().getSRatedAC)
-        flexResult.getpMin should beEquivalentTo(ev900.unwrap().getSRatedAC)
-        flexResult.getpMax should beEquivalentTo(ev900.unwrap().getSRatedAC)
+        flexResult.getpRef should beEquivalentTo(ev900.unwrap().getPRatedAC)
+        flexResult.getpMin should beEquivalentTo(ev900.unwrap().getPRatedAC)
+        flexResult.getpMax should beEquivalentTo(ev900.unwrap().getPRatedAC)
       }
 
       emAgent.send(evcsAgent, IssueNoControl(900))
@@ -1397,7 +1397,7 @@ class EvcsAgentModelCalculationSpec
           Kilowatts(
             ev900
               .unwrap()
-              .getSRatedAC
+              .getPRatedAC
               .to(PowerSystemUnits.KILOWATT)
               .getValue
               .doubleValue
@@ -1451,7 +1451,7 @@ class EvcsAgentModelCalculationSpec
           case ParticipantResultEvent(result: EvResult)
               if result.getTime.equals(900.toDateTime) =>
             result.getInputModel shouldBe ev900.uuid
-            result.getP should beEquivalentTo(ev900.unwrap().getSRatedAC)
+            result.getP should beEquivalentTo(ev900.unwrap().getPRatedAC)
             result.getQ should beEquivalentTo(0d.asMegaVar)
             result.getSoc should beEquivalentTo(0d.asPercent)
           case ParticipantResultEvent(result: EvResult)
@@ -1466,7 +1466,7 @@ class EvcsAgentModelCalculationSpec
         case ParticipantResultEvent(result: EvcsResult) =>
           result.getInputModel shouldBe evcsInputModelQv.getUuid
           result.getTime shouldBe 900.toDateTime
-          result.getP should beEquivalentTo(ev900.unwrap().getSRatedAC)
+          result.getP should beEquivalentTo(ev900.unwrap().getPRatedAC)
           result.getQ should beEquivalentTo(0d.asMegaVar)
       }
 
@@ -1500,9 +1500,9 @@ class EvcsAgentModelCalculationSpec
       resultListener.expectMsgPF() { case FlexOptionsResultEvent(flexResult) =>
         flexResult.getInputModel shouldBe evcsInputModelQv.getUuid
         flexResult.getTime shouldBe 4500.toDateTime
-        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getSRatedAC)
-        flexResult.getpMin should beEquivalentTo(ev4500.unwrap().getSRatedAC)
-        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getSRatedAC)
+        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getPRatedAC)
+        flexResult.getpMin should beEquivalentTo(ev4500.unwrap().getPRatedAC)
+        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getPRatedAC)
       }
 
       emAgent.send(evcsAgent, IssueNoControl(4500))
@@ -1550,9 +1550,9 @@ class EvcsAgentModelCalculationSpec
       resultListener.expectMsgPF() { case FlexOptionsResultEvent(flexResult) =>
         flexResult.getInputModel shouldBe evcsInputModelQv.getUuid
         flexResult.getTime shouldBe 9736.toDateTime
-        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getSRatedAC)
+        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getPRatedAC)
         flexResult.getpMin should beEquivalentTo(0d.asKiloWatt)
-        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getSRatedAC)
+        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getPRatedAC)
       }
 
       emAgent.send(evcsAgent, IssuePowerControl(9736, Kilowatts(10.0)))
@@ -1615,7 +1615,7 @@ class EvcsAgentModelCalculationSpec
       emAgent.send(evcsAgent, FlexActivation(11700))
 
       val combinedChargingPower =
-        ev11700.unwrap().getSRatedAC.add(ev4500.unwrap().getSRatedAC)
+        ev11700.unwrap().getPRatedAC.add(ev4500.unwrap().getPRatedAC)
       val combinedChargingPowerSq = Kilowatts(
         combinedChargingPower.to(PowerSystemUnits.KILOWATT).getValue.doubleValue
       )
@@ -1640,7 +1640,7 @@ class EvcsAgentModelCalculationSpec
         flexResult.getTime shouldBe 11700.toDateTime
         flexResult.getpRef should beEquivalentTo(combinedChargingPower)
         flexResult.getpMin should beEquivalentTo(
-          ev4500.unwrap().getSRatedAC.multiply(-1)
+          ev4500.unwrap().getPRatedAC.multiply(-1)
         )
         flexResult.getpMax should beEquivalentTo(combinedChargingPower)
       }
@@ -1792,7 +1792,7 @@ class EvcsAgentModelCalculationSpec
         flexResult.getpMin should beEquivalentTo(
           ev4500
             .unwrap()
-            .getSRatedAC
+            .getPRatedAC
             .multiply(
               -1
             )
@@ -1985,9 +1985,9 @@ class EvcsAgentModelCalculationSpec
       resultListener.expectMsgPF() { case FlexOptionsResultEvent(flexResult) =>
         flexResult.getInputModel shouldBe evcsInputModelQv.getUuid
         flexResult.getTime shouldBe 36000.toDateTime
-        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getSRatedAC)
+        flexResult.getpRef should beEquivalentTo(ev4500.unwrap().getPRatedAC)
         flexResult.getpMin should beEquivalentTo(0.asKiloWatt)
-        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getSRatedAC)
+        flexResult.getpMax should beEquivalentTo(ev4500.unwrap().getPRatedAC)
       }
 
       emAgent.send(evcsAgent, IssuePowerControl(36000, Kilowatts(4.0)))